fix: restore explicit types dropped in unrelated CLion edits
clusteredlights.cpp, model.cpp, and shadowmap.cpp carried unrelated
type-explicitness regressions from an earlier out-of-band edit
(dropped static_cast<uint32_t> on a dispatch() arg, CTAD replacing
explicit std::vector<glm::vec3> and std::array<float, N>) that got
swept into the previous commit because those files were also touched
for lint fixes. Restore the explicit types; no behavior change.

+++ sponge/src/platform/opengl/scene/clusteredlights.cpp
@@ -144,7 +144,7 @@ void ClusteredLights::update(const glm::vec3* positions,
     clusterAABBsSSBO.bindBase(6);
     computeParamsSSBO.bindBase(7);
 
-    assignShader.dispatch((maxClusters + 63) / 64);
+    assignShader.dispatch(static_cast<uint32_t>((maxClusters + 63) / 64));
 
     glMemoryBarrier(GL_SHADER_STORAGE_BARRIER_BIT);
 }
